Add browserId in email submission form
