+++ dotcom-rendering/src/components/Metrics.importable.tsx
@@ -9,8 +9,9 @@ import {
 	bypassCoreWebVitalsSampling,
 	initCoreWebVitals,
 } from '@guardian/core-web-vitals';
-import { getCookie, isString, isUndefined } from '@guardian/libs';
+import { isUndefined } from '@guardian/libs';
 import { useCallback, useEffect, useState } from 'react';
+import { useBrowserId } from '../lib/metrics';
 import { useAB, useBetaAB } from '../lib/useAB';
 import { useAdBlockInUse } from '../lib/useAdBlockInUse';
 import { useDetectAdBlock } from '../lib/useDetectAdBlock';
@@ -39,20 +40,6 @@ const shouldCollectMetricsForBetaTests = (userTestParticipations: string[]) => {
 	return userParticipationConfigs.some(
 		(test) => test.shouldForceMetricsCollection,
 	);
-};
-
-const useBrowserId = () => {
-	const [browserId, setBrowserId] = useState<string>();
-
-	useEffect(() => {
-		const cookie = getCookie({ name: 'bwid', shouldMemoize: true });
-
-		const id = isString(cookie) ? cookie : 'no-browser-id-available';
-
-		setBrowserId(id);
-	}, []);
-
-	return browserId;
 };
 
 const useDev = () => {
